+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.11.6] - 2025-10-10
+
+### Added
+
+- All IndexTTS-2 dependencies now download to TTS folder as intended
+
+### Changed
+
+- Better organization of model files following project standards
+
+### Fixed
+
+- Fix IndexTTS-2 model download policy violations
+- Fix w2v-bert-2.0 model downloading to wrong directory
+- Fix MaskGCT semantic codec downloading to wrong directory
+- Fix BigVGAN vocoder downloading to wrong directory
 ## [4.11.5] - 2025-10-10
 
 ### Added